stop removal of admin user

+++ com.mobi.jaas.rest/src/main/java/com/mobi/jaas/rest/UserRest.java
@@ -39,6 +39,7 @@ import com.mobi.jaas.api.ontologies.usermanagement.Role;
 import com.mobi.jaas.api.ontologies.usermanagement.User;
 import com.mobi.jaas.api.ontologies.usermanagement.UserFactory;
 import com.mobi.persistence.utils.api.SesameTransformer;
+import com.mobi.rdf.api.IRI;
 import com.mobi.rdf.api.Model;
 import com.mobi.rdf.api.Resource;
 import com.mobi.rdf.api.Value;
@@ -88,6 +89,7 @@ public class UserRest {
     private SesameTransformer transformer;
     private Engine rdfEngine;
     private final Logger logger = LoggerFactory.getLogger(UserRest.class);
+    private static final String ADMIN_USER_IRI = "http://mobi.com/users/d033e22ae348aeb5660fc2140aec35850c4da997";
 
     @Reference
     void setEngineManager(EngineManager engineManager) {
@@ -380,7 +382,10 @@ public class UserRest {
             if (!engineManager.userExists(username)) {
                 throw ErrorUtils.sendError("User " + username + " not found", Response.Status.BAD_REQUEST);
             }
-
+            IRI userIRI = (IRI) engineManager.retrieveUser(username).get().getResource();
+            if (userIRI.equals(vf.createIRI(ADMIN_USER_IRI))) {
+                throw ErrorUtils.sendError("The admin user cannot be deleted.", Response.Status.METHOD_NOT_ALLOWED);
+            }
             engineManager.deleteUser(rdfEngine.getEngineName(), username);
             logger.info("Deleted user " + username);
             return Response.ok().build();
